+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -334,7 +334,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KX2XC0QC7EP673CNWSD6BNJW"
+        "01KX3D75MG304H8AVYPDKHEC4D"
     )
     r"""The unique identifier of the data source"""