+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KX2XC0QGPYXQ1J2K2JCZQVYC"
+        "01KX3D75MMN0T6H9XJXN3S6834"
     )
     r"""The unique identifier of the data source"""