+++ packages/orq-rc/RELEASES.md
@@ -9379,3 +9379,13 @@ Based on:
 - [python v4.12.0-rc.10] packages/orq-rc
 ### Releases
 - [PyPI v4.12.0-rc.10] https://pypi.org/project/orq-ai-sdk/4.12.0-rc.10 - packages/orq-rc
+
+## 2026-07-09 12:24:20
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.789.0 (2.916.2) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.12.0-rc.11] packages/orq-rc
+### Releases
+- [PyPI v4.12.0-rc.11] https://pypi.org/project/orq-ai-sdk/4.12.0-rc.11 - packages/orq-rc